Remove redis
If redis isn't used, removing it here helps understand the barrier to setup much better. I thought I had to setup a redis cluster to support Bullet Train, but it looks like that isn't the case (I can't see any reference to redis in the code other than in this file, but I may be missing something.